filter/filter.go
package filter

import (
	"regexp"

	pivnet "github.com/pivotal-cf/go-pivnet"
)

type Filter struct {
}

func NewFilter() *Filter {
	return &Filter{}
}

// ReleasesByVersion returns all releases that match the provided version regex
func (f Filter) ReleasesByVersion(releases []pivnet.Release, version string) ([]pivnet.Release, error) {
	filteredReleases := make([]pivnet.Release, 0)

	for _, release := range releases {
		match, err := regexp.MatchString(version, release.Version)
		if err != nil {
			return nil, err
		}

		if match {
			filteredReleases = append(filteredReleases, release)
		}
	}

	return filteredReleases, nil
}
